Improved locking mechanics by getting rid of synchronized blocks.
Instead of using synchronized blocks on different parts of internal data
structures, I decided to use a ReadWriteLock for much clearer locking
mechanics. The reason for this decision is the fact that we have to
transfer values to raw and filtered event buffers when flushing from
work buffer, but we only read from filtered event buffer when drawing.
This would nested synchronized blocks which are way less readable.